+++ app/src/main/AndroidManifest.xml
@@ -15,12 +15,15 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
         android:usesCleartextTraffic="true">
+        <activity
+            android:name=".ui.settings.SettingsActivity"
+            android:label="@string/settings"></activity>
 
         <meta-data
             android:name="com.kakao.sdk.AppKey"
             android:value="397f6605753c6e11dd0484f23dcba8ca" />
 
-        <activity android:name=".ui.details.LineDetailsActivity"></activity>
+        <activity android:name=".ui.details.LineDetailsActivity" />
         <activity android:name=".ui.splash.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -28,7 +31,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.MainActivity"></activity>
+        <activity android:name=".ui.MainActivity" />
 
         <receiver
             android:name=".receiver.AlarmReceiver"
@@ -38,12 +41,10 @@
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
             </intent-filter>
         </receiver>
-
         <receiver
             android:name=".receiver.GeofenceBroadcastReceiver"
             android:enabled="true"
-            android:exported="false"></receiver>
-
+            android:exported="false" />
     </application>
 
 </manifest>
