+++ app/src/main/res/navigation/nav_graph.xml
@@ -20,9 +20,4 @@
         android:name="com.hppk.sw.hppkcommuterbus.ui.mypage.MyPageFragment"
         android:label="fragment_my_page"
         tools:layout="@layout/fragment_my_page" />
-    <fragment
-        android:id="@+id/settingsFragment"
-        android:name="com.hppk.sw.hppkcommuterbus.ui.settings.SettingsFragment"
-        android:label="fragment_settings"
-        tools:layout="@layout/fragment_settings" />
 </navigation>
